+++ api/utils/validator.js
@@ -199,7 +199,6 @@ const sanitizzaStringa = (input) => {
         .replace(/>/g, '&gt;')
         .replace(/"/g, '&quot;')
         .replace(/'/g, '&#x27;')
-        .replace(/\//g, '&#x2F;')
         .trim();
 };